+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## Unreleased
+
+### Added
+
+- Link settings that enable registrations and invites
+
+### Fixed
+
+- Disable Invites tab when invites are disabled on BE
+
 ## [2.0.2] - 2020-04-01
 
 ### Added