+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+## [2.0.3] - 2020-04-29
+
 ### Added
 
 - Link settings that enable registrations and invites